requirements.txt
ccxt>=2.0.0
pandas>=1.5.0
requests>=2.28.0
python-telegram-bot>=13.0
numpy>=1.21.0
scikit-learn>=1.0.0
python-dotenv>=0.19.0
schedule>=1.1.0
streamlit>=1.24.0
plotly>=5.13.0
altair>=5.0.0